Revert "Revert "Rename "Wakeable" modules to "Thenable"""
This reverts commit 36aba8fb6301915974569c16132dc683820cff51.

This reverts commit 42aff59d42b4c9b61b8e228d8b373aef976f47b9.

The distinction between "wakeable" and "thenable" isn't really
important in these files anymore so we should just pick one.
